Moved into common/application
