+++ src/test/activation/activationManager.unit.test.ts
@@ -9,10 +9,10 @@ import * as typemoq from 'typemoq';
 import { TextDocument, Uri } from 'vscode';
 import { ExtensionActivationManager } from '../../client/activation/activationManager';
 import { LanguageServerExtensionActivationService } from '../../client/activation/activationService';
-import { ActiveResourceService } from '../../client/activation/activeResource';
-import { IActiveResourceService, IExtensionActivationService } from '../../client/activation/types';
+import { IExtensionActivationService } from '../../client/activation/types';
 import { IApplicationDiagnostics } from '../../client/application/types';
-import { IDocumentManager, IWorkspaceService } from '../../client/common/application/types';
+import { ActiveResourceService } from '../../client/common/application/activeResource';
+import { IActiveResourceService, IDocumentManager, IWorkspaceService } from '../../client/common/application/types';
 import { WorkspaceService } from '../../client/common/application/workspace';
 import { PYTHON_LANGUAGE } from '../../client/common/constants';
 import { IDisposable } from '../../client/common/types';
