+++ src/test/common/terminals/activation.unit.test.ts
@@ -6,10 +6,9 @@ import { expect } from 'chai';
 import { anything, instance, mock, verify, when } from 'ts-mockito';
 import * as TypeMoq from 'typemoq';
 import { Terminal, Uri } from 'vscode';
-import { ActiveResourceService } from '../../../client/activation/activeResource';
-import { IActiveResourceService } from '../../../client/activation/types';
+import { ActiveResourceService } from '../../../client/common/application/activeResource';
 import { TerminalManager } from '../../../client/common/application/terminalManager';
-import { ITerminalManager } from '../../../client/common/application/types';
+import { IActiveResourceService, ITerminalManager } from '../../../client/common/application/types';
 import { TerminalActivator } from '../../../client/common/terminal/activator';
 import { ITerminalActivator } from '../../../client/common/terminal/types';
 import { IDisposable } from '../../../client/common/types';
