+++ src/test/providers/terminal.unit.test.ts
@@ -5,8 +5,7 @@ import * as assert from 'assert';
 import { expect } from 'chai';
 import * as TypeMoq from 'typemoq';
 import { Disposable, Terminal, Uri } from 'vscode';
-import { IActiveResourceService } from '../../client/activation/types';
-import { ICommandManager, IWorkspaceService } from '../../client/common/application/types';
+import { IActiveResourceService, ICommandManager, IWorkspaceService } from '../../client/common/application/types';
 import { Commands } from '../../client/common/constants';
 import { TerminalService } from '../../client/common/terminal/service';
 import { ITerminalActivator, ITerminalServiceFactory } from '../../client/common/terminal/types';
